+++ messenger-core/src/main/java/com/example/messenger/security/jwt/JwtTokenUtil.java
@@ -1,4 +1,4 @@
-package com.example.messenger.jwt;
+package com.example.messenger.security.jwt;
 
 import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.Jwts;
